+++ packages/moonstone/Scroller/Scroller.js
@@ -9,6 +9,7 @@
 import classNames from 'classnames';
 import {contextTypes} from '@enact/i18n/I18nDecorator';
 import deprecate from '@enact/core/internal/deprecate';
+import {forward} from '@enact/core/handle';
 import {getTargetByDirectionFromElement, getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
@@ -17,9 +18,11 @@ import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDeco
 
 import css from './Scroller.less';
 import Scrollable from './Scrollable';
+import {RemeasurableDecorator} from '@enact/ui/Remeasurable';
 
 const
 	dataContainerDisabledAttribute = 'data-container-disabled',
+	epsilon = 1,
 	reverseDirections = {
 		'left': 'right',
 		'right': 'left'
@@ -236,11 +239,12 @@ class ScrollerBase extends Component {
 	 * @param {Node} item node
 	 * @param {Object} scrollInfo position info. `calculateScrollTop` uses
 	 * `scrollInfo.previousScrollHeight` and `scrollInfo.scrollTop`
+	 * @param {Number} scrollPosition last target position, passed scroll animation is ongoing
 	 *
 	 * @returns {Object} with keys {top, left} containing caculated top and left positions for scroll.
 	 * @private
 	 */
-	calculatePositionOnFocus = ({item, scrollInfo}) => {
+	calculatePositionOnFocus = ({item, scrollInfo, scrollPosition}) => {
 		if (!this.isVertical() && !this.isHorizontal() || !item || !this.containerRef.contains(item)) {
 			return;
 		}
@@ -253,7 +257,7 @@ class ScrollerBase extends Component {
 		} = this.getFocusedItemBounds(item);
 
 		if (this.isVertical()) {
-			this.scrollPos.top = this.calculateScrollTop(item, itemTop, itemHeight, scrollInfo);
+			this.scrollPos.top = this.calculateScrollTop(item, itemTop, itemHeight, scrollInfo, scrollPosition);
 		}
 
 		if (this.isHorizontal()) {
@@ -261,7 +265,8 @@ class ScrollerBase extends Component {
 				{clientWidth} = this.scrollBounds,
 				rtlDirection = this.context.rtl ? -1 : 1,
 				{left: containerLeft} = this.containerRef.getBoundingClientRect(),
-				currentScrollLeft = this.context.rtl ? (this.scrollBounds.maxLeft - this.scrollPos.left) : this.scrollPos.left,
+				scrollLastPosition = scrollPosition ? scrollPosition : this.scrollPos.left,
+				currentScrollLeft = this.context.rtl ? (this.scrollBounds.maxLeft - scrollLastPosition) : scrollLastPosition,
 				// calculation based on client position
 				newItemLeft = this.containerRef.scrollLeft + (itemLeft - containerLeft);
 
@@ -287,15 +292,16 @@ class ScrollerBase extends Component {
 	 * @param {Number} itemHeight of focusedItem / focusedContainer
 	 * @param {Object} scrollInfo position info. Uses `scrollInfo.previousScrollHeight`
 	 * and `scrollInfo.scrollTop`
+	 * @param {Number} scrollPosition last target position, passed scroll animation is ongoing
 	 *
 	 * @returns {Number} Calculated `scrollTop`
 	 * @private
 	 */
-	calculateScrollTop = (focusedItem, itemTop, itemHeight, scrollInfo) => {
+	calculateScrollTop = (focusedItem, itemTop, itemHeight, scrollInfo, scrollPosition) => {
 		const
 			{clientHeight} = this.scrollBounds,
 			{top: containerTop} = this.containerRef.getBoundingClientRect(),
-			currentScrollTop = this.scrollPos.top,
+			currentScrollTop = (scrollPosition ? scrollPosition : this.scrollPos.top),
 			// calculation based on client position
 			newItemTop = this.containerRef.scrollTop + (itemTop - containerTop),
 			itemBottom = newItemTop + itemHeight,
@@ -344,20 +350,20 @@ class ScrollerBase extends Component {
 				nestedItemTop = this.containerRef.scrollTop + (top - containerTop),
 				nestedItemBottom = nestedItemTop + nestedItemHeight;
 
-			if (newItemTop - nestedItemHeight > currentScrollTop) {
+			if (newItemTop - nestedItemHeight - currentScrollTop > epsilon) {
 				// set scroll position so that the top of the container is at least on the top
 				newScrollTop = newItemTop - nestedItemHeight;
-			} else if (nestedItemBottom > scrollBottom) {
+			} else if (nestedItemBottom - scrollBottom > epsilon) {
 				// Caculate when 5-way focus down past the bottom.
 				newScrollTop += nestedItemBottom - scrollBottom;
-			} else if (nestedItemTop < currentScrollTop) {
+			} else if (nestedItemTop - currentScrollTop < epsilon) {
 				// Caculate when 5-way focus up past the top.
 				newScrollTop += nestedItemTop - currentScrollTop;
 			}
-		} else if (itemBottom > scrollBottom) {
+		} else if (itemBottom - scrollBottom > epsilon) {
 			// Caculate when 5-way focus down past the bottom.
 			newScrollTop += itemBottom - scrollBottom;
-		} else if (newItemTop < currentScrollTop) {
+		} else if (newItemTop - currentScrollTop < epsilon) {
 			// Caculate when 5-way focus up past the top.
 			newScrollTop += newItemTop - currentScrollTop;
 		}
@@ -413,7 +419,10 @@ class ScrollerBase extends Component {
 		scrollBounds.maxTop = Math.max(0, scrollHeight - clientHeight);
 	}
 
-	onKeyDown = ({keyCode, target}) => {
+	onKeyDown = (ev) => {
+		forward('onKeyDown', ev, this.props);
+
+		const {keyCode, target} = ev;
 		const direction = getDirection(keyCode);
 
 		if (direction && !this.findInternalTarget(direction, target)) {
@@ -477,8 +486,11 @@ class ScrollerBase extends Component {
  */
 const Scroller = SpotlightContainerDecorator(
 	{restrict: 'self-first'},
-	Scrollable(
-		ScrollerBase
+	RemeasurableDecorator(
+		{trigger: 'onScrollbarVisibilityChange'},
+		Scrollable(
+			ScrollerBase
+		)
 	)
 );