+++ packages/sampler/stories/moonstone-stories/Marquee.js
@@ -20,7 +20,7 @@ storiesOf('Marquee')
 						marqueeDelay={number('marqueeDelay', 1000)}
 						marqueeDisabled={nullify(boolean('marqueeDisabled', false))}
 						marqueeOn={select('marqueeOn', ['hover', 'render'], 'render')}
-						marqueeOnRenderDelay={number('marqueeOnRenderDelay', 1000)}
+						marqueeOnRenderDelay={1000}
 						marqueeResetDelay={number('marqueeResetDelay', 1000)}
 						marqueeSpeed={number('marqueeSpeed', 60)}
 						style={{width: '400px'}}