+++ packages/sampler/stories/qa-stories/Remeasurable.js
@@ -5,27 +5,40 @@ import React from 'react';
 import {storiesOf} from '@kadira/storybook';
 import SelectableItem from '@enact/moonstone/SelectableItem';
 
-
 const data = [
 	'a',
 	'ABCDEFGHIJKLMNOPQRSTUVW',
 	'c'
 ];
 
+class NoUpdate extends React.Component {
+	shouldComponentUpdate () {
+		return false;
+	}
+
+	render () {
+		return (
+			<div>{this.props.children}</div>
+		);
+	}
+}
+
 storiesOf('Remeasurable')
 	.addWithInfo(
 		'should recalculate long marquee when scrollbar is rendered',
 		() => (
 			<Scroller style={{height: '400px', width: '500px'}}>
-				<Item>ABCDEFGHIJKLMNOPQRST</Item>
-				<SelectableItem>
-					SELECTABLE ITEM ABCDEFG
-				</SelectableItem>
-				<ExpandableList title={'ABCDEFGHIJKLMNOPQRS'}>
-					{data}
-				</ExpandableList>
-				<Item>dummy</Item>
-				<Item>dummy</Item>
+				<NoUpdate>
+					<Item>ABCDEFGHIJKLMNOPQRST</Item>
+					<SelectableItem>
+						SELECTABLE ITEM ABCDEFG
+					</SelectableItem>
+					<ExpandableList title={'ABCDEFGHIJKLMNOPQRS'}>
+						{data}
+					</ExpandableList>
+					<Item>dummy</Item>
+					<Item>dummy</Item>
+				</NoUpdate>
 			</Scroller>
 		)
 	)